refactor(web): HTTPエラー写像を各リソースへコロケートする
エラー写像を server/error 集約から各リソース配下(article/passkey/
sessions/registrations)へ移し、写像先を持つエラーだけをリソース内の
テーブルで定義する。写像対象を持たない oauth 開始・ログアウトはサーバ
起因の失敗として errorHandler の 5xx 処理へ委ねる。
