--- application/apps/web/src/server/error/auth-error.ts
@@ -1,16 +1,0 @@
-import {
-  InvalidCredentialsError,
-  NoSessionError,
-  UserAlreadyExistsError,
-} from '@trend-diary/authentication'
-import throwHttpErrorByTable, { type ErrorStatusTable } from './throw-http-error'
-
-const ERROR_STATUS_TABLE: ErrorStatusTable = [
-  [InvalidCredentialsError, 401],
-  [UserAlreadyExistsError, 409],
-  [NoSessionError, 401],
-]
-
-export default function throwHttpError(error: Error): never {
-  return throwHttpErrorByTable(error, ERROR_STATUS_TABLE)
-}
